@@ -4,7 +4,7 @@ project(Project VERSION 1.0 LANGUAGES CXX)
 
 add_subdirectory(dependencies)
 
-set(CMAKE_CXX_STANDARD 17)
+set(CMAKE_CXX_STANDARD 23)
 set(CMAKE_CXX_STANDARD_REQUIRED TRUE)
 set(CMAKE_EXPORT_COMPILE_COMMANDS ON)
 
