@@ -1,21 +1,40 @@
-#include "capd/capdlib.h"
+#include <print>
+
+#include <capd/capdlib.h>
 
 #include "common.h"
 
-static long double getAngle(long double pos1[2], long double pos2[2]) {
-    long double x1 = pos1[0];
-    long double x2 = pos2[0];
-    long double y1 = pos1[1];
-    long double y2 = pos2[1];
-
-    long double r1 = sqrt(x1 * x1 + y1 * y1);
-    long double r2 = sqrt(x2 * x2 + y2 * y2);
-
-    return acos((x1 * x2 + y1 * y2) / (r1 * r2));
+num toDeg(num rad) {
+    return 180 * rad / M_PI;
 }
 
+static num getAngle(num pos1[2], num pos2[2]) {
+    num x1 = pos1[0];
+    num x2 = pos2[0];
+    num y1 = pos1[1];
+    num y2 = pos2[1];
+
+    num r1 = sqrt(x1 * x1 + y1 * y1);
+    num r2 = sqrt(x2 * x2 + y2 * y2);
+
+    num rad = acos((x1 * x2 + y1 * y2) / (r1 * r2));
+
+    return toDeg(rad);
+}
+
+typedef capd::vectalg::Matrix<num, 0, 0> Matrix;
+typedef capd::map::Map<Matrix> Map;
+typedef capd::dynsys::BasicOdeSolver<Map> Solver;
+typedef capd::vectalg::Vector<num, 0> Vector;
+
+struct timestamp {
+    Vector pos;
+    num r;
+    num t;
+};
+
 void simulate(struct commonData &data) {
-    capd::LDMap pendulum(
+    Map pendulum(
         "time:t;"
         "var:"
             "x,"
@@ -35,39 +54,53 @@ void simulate(struct commonData &data) {
     pendulum.setParameter("cs", 0.00001);
 
     uint32_t order = 100;
-    capd::LDOdeSolver solver{pendulum, order};
+
+    Solver solver{pendulum, order};
 
     solver.setStep(0.00001);
 
-    capd::LDVector u{1.0, 0.0, 0.0, 0.06};
-    long double t = 0.0;
+    Vector curr{1.0, 0.0, 0.0, data.init};
+    num t = 0.0;
 
-    size_t cooldown = 0;
-    long double max = 0;
-    long double prev[2] = { u[0], u[1] };
+    struct timestamp max = {
+        .pos = curr,
+        .r = 1
+    };
+
+    struct timestamp last[3] = { 
+        { curr, 1, 0 }, 
+        { curr, 1, 0 }, 
+        { curr, 1, 0 }, 
+    };
 
     std::vector<struct data> temp;
 
     do {
-        u = solver(t, u);
+        curr = solver(t, curr);
 
-        long double curr = sqrt(u[0] * u[0] + u[1] * u[1]);
+        last[2] = last[1];
+        last[1] = last[0];
+        last[0] = {
+            .pos = curr,
+            .r = sqrt(curr[0] * curr[0] + curr[1] * curr[1]),
+            .t = t
+        };
+
         temp.emplace_back((struct data) {
             .t = t,
-            .pos = { u[0], u[1] }
+            .pos = { curr[0], curr[1] }
         });
+ 
+        if (last[1].r >= last[2].r && last[1].r >= last[0].r) {
+            std::print("\33[2K\rMax Detected at {} - {:.6}\n", last[1].t, getAngle(
+                (num[]) { max.pos[0], max.pos[1] }, 
+                (num[]) { last[1].pos[0], last[1].pos[1] }
+            ));
 
-        if (cooldown > 0) cooldown -= 1;
-
-        if (max < curr || abs(max - curr) < 10e-10) 
-        if (cooldown == 0) {
-            long double uu[2] = { u[0], u[1] };
-
-            printf("\n%Lf\n", getAngle(prev, uu));
-            max = curr;
-            prev[0] = u[0];
-            prev[1] = u[1];
-            cooldown = 50;
+            max = {
+                .pos = last[1].pos,
+                .r = std::max(last[1].r, max.r)
+            };
         }
 
         if (data.isDrawOrdered == false && data.access.try_lock()) {
@@ -81,8 +114,7 @@ void simulate(struct commonData &data) {
 
             data.drawOrderMessager.notify_all();
         }
-
-        printf("\rt=%6Lf, x=%6Lf, y=%6Lf, dx=%6Lf, dy=%6Lf", t, u[0], u[1], u[2], u[3]);
+        std::print("\r{:.6f}: x={:.6f}, y={:.6f}, dx={:.6f}, dy={:.6f}, max={:.6f}, curr={:.6f}", t, curr[0], curr[1], curr[2], curr[3], max.r, last[0].r);
     } while(t < 20'000);
 
     data.access.lock();
